Refactoring: new data approach, new items data

+++ UnityDeveloperTestTask-zdoveloV/Assets/Client/Code/Game/UI/Elements/Currency/CurrencyViewData.cs
@@ -6,7 +6,7 @@ namespace Client.Code.Game.UI.Elements.Currency
     [Serializable]
     public class CurrencyViewData
     {
-        public int Value;
+        public long Value;
         public Sprite Icon;
     }
 }
